+++ src/main/java/com/feiyu/discord/sg/tavern/schedulers/IntroMsgScheduler.java
@@ -90,10 +90,10 @@ public class IntroMsgScheduler {
         log.info("Message to be sent : {}", message.replaceAll("\\n", ", "));
         
         //adminChannel.sendMessage(message).complete();
-        Member adminMember = guild.retrieveMemberById(valuesConfig.getAdminUserId()).complete();
-        log.info("Message sent to : {}", adminMember);
-        PrivateChannel pc = adminMember.getUser().openPrivateChannel().complete();
-        pc.sendMessage(message).queue();
+//        Member adminMember = guild.retrieveMemberById(valuesConfig.getAdminUserId()).complete();
+//        log.info("Message sent to : {}", adminMember);
+//        PrivateChannel pc = adminMember.getUser().openPrivateChannel().complete();
+//        pc.sendMessage(message).queue();
         log.info("IntroMsgScheduler.introReminder End");
     }
 }